Move ExpiredTokenException to the correct namespace
